Uninter 2/requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0
lxml==4.9.3
webdriver-manager==4.0.1
gunicorn==21.2.0
